+++ engines/twine/scene/scene.cpp
@@ -153,8 +153,8 @@ bool Scene::loadSceneLBA2() {
 	_currentGameOverScene = stream.readByte();
 	stream.skip(4);
 
-	_alphaLight = ClampAngle(stream.readUint16LE());
-	_betaLight = ClampAngle(stream.readUint16LE());
+	_alphaLight = ClampAngle((int16)stream.readUint16LE());
+	_betaLight = ClampAngle((int16)stream.readUint16LE());
 	debug(2, "Using %i and %i as light vectors", _alphaLight, _betaLight);
 
 	_isOutsideScene = stream.readByte();
@@ -177,34 +177,34 @@ bool Scene::loadSceneLBA2() {
 	_sceneHeroPos.y = stream.readSint16LE();
 	_sceneHeroPos.z = stream.readSint16LE();
 
-	_sceneHero->_moveScriptSize = stream.readUint16LE();
+	_sceneHero->_moveScriptSize = (int16)stream.readUint16LE();
 	_sceneHero->_moveScript = _currentScene + stream.pos();
 	stream.skip(_sceneHero->_moveScriptSize);
 
-	_sceneHero->_lifeScriptSize = stream.readUint16LE();
+	_sceneHero->_lifeScriptSize = (int16)stream.readUint16LE();
 	_sceneHero->_lifeScript = _currentScene + stream.pos();
 	stream.skip(_sceneHero->_lifeScriptSize);
 
-	_sceneNumActors = stream.readUint16LE();
+	_sceneNumActors = (int16)stream.readUint16LE();
 	int cnt = 1;
 	for (int32 a = 1; a < _sceneNumActors; a++, cnt++) {
 		_engine->_actor->resetActor(a);
 		ActorStruct *act = &_sceneActors[a];
 		setActorStaticFlags(act, stream.readUint32LE());
 
-		act->loadModel(stream.readUint16LE(), false);
+		act->loadModel((int16)stream.readUint16LE(), false);
 
 		act->_body = (BodyType)stream.readSint16LE();
 		act->_anim = (AnimationTypes)stream.readByte();
-		act->_sprite = stream.readUint16LE();
-		act->_pos.x = stream.readUint16LE();
-		act->_pos.y = stream.readUint16LE();
-		act->_pos.z = stream.readUint16LE();
+		act->_sprite = (int16)stream.readUint16LE();
+		act->_pos.x = (int16)stream.readUint16LE();
+		act->_pos.y = (int16)stream.readUint16LE();
+		act->_pos.z = (int16)stream.readUint16LE();
 		act->_collisionPos = act->pos();
 		act->_strengthOfHit = stream.readByte();
 		setBonusParameterFlags(act, stream.readUint16LE());
-		act->_angle = stream.readUint16LE();
-		act->_speed = stream.readUint16LE();
+		act->_angle = (int16)stream.readUint16LE();
+		act->_speed = (int16)stream.readUint16LE();
 		act->_controlMode = (ControlMode)stream.readByte();
 		act->_cropLeft = stream.readSint16LE();
 		act->_delayInMillis = act->_cropLeft; // TODO: this might not be needed
@@ -222,11 +222,11 @@ bool Scene::loadSceneLBA2() {
 		act->_armor = stream.readByte();
 		act->setLife(stream.readByte());
 
-		act->_moveScriptSize = stream.readUint16LE();
+		act->_moveScriptSize = (int16)stream.readUint16LE();
 		act->_moveScript = _currentScene + stream.pos();
 		stream.skip(act->_moveScriptSize);
 
-		act->_lifeScriptSize = stream.readUint16LE();
+		act->_lifeScriptSize = (int16)stream.readUint16LE();
 		act->_lifeScript = _currentScene + stream.pos();
 		stream.skip(act->_lifeScriptSize);
 
@@ -236,7 +236,7 @@ bool Scene::loadSceneLBA2() {
 		}
 	}
 
-	_sceneNumZones = stream.readUint16LE();
+	_sceneNumZones = (int16)stream.readUint16LE();
 	for (int32 i = 0; i < _sceneNumZones; i++) {
 		ZoneStruct *zone = &_sceneZones[i];
 		zone->mins.x = stream.readSint32LE();
@@ -260,7 +260,7 @@ bool Scene::loadSceneLBA2() {
 		zone->snap = stream.readUint16LE();
 	}
 
-	_sceneNumTracks = stream.readUint16LE();
+	_sceneNumTracks = (int16)stream.readUint16LE();
 	for (int32 i = 0; i < _sceneNumTracks; i++) {
 		IVec3 *point = &_sceneTracks[i];
 		point->x = stream.readSint32LE();
@@ -268,8 +268,8 @@ bool Scene::loadSceneLBA2() {
 		point->z = stream.readSint32LE();
 	}
 
-	int32 sceneNumPatches = stream.readUint16LE();
-	for (int32 i = 0; i < sceneNumPatches; i++) {
+	uint16 sceneNumPatches = stream.readUint16LE();
+	for (uint16 i = 0; i < sceneNumPatches; i++) {
 		/*size = */stream.readUint16LE();
 		/*offset = */stream.readUint16LE();
 	}
@@ -287,8 +287,8 @@ bool Scene::loadSceneLBA1() {
 
 	// FIXME: Workaround to fix lighting issue - not using proper dark light
 	// Using 1215 and 1087 as light vectors - scene 8
-	_alphaLight = ClampAngle(stream.readUint16LE());
-	_betaLight = ClampAngle(stream.readUint16LE());
+	_alphaLight = ClampAngle((int16)stream.readUint16LE());
+	_betaLight = ClampAngle((int16)stream.readUint16LE());
 	debug(2, "Using %i and %i as light vectors", _alphaLight, _betaLight);
 
 	for (int i = 0; i < 4; ++i) {
@@ -315,7 +315,7 @@ bool Scene::loadSceneLBA1() {
 	_sceneHero->_lifeScript = _currentScene + stream.pos();
 	stream.skip(_sceneHero->_lifeScriptSize);
 
-	_sceneNumActors = stream.readUint16LE();
+	_sceneNumActors = (int16)stream.readUint16LE();
 	int cnt = 1;
 	for (int32 a = 1; a < _sceneNumActors; a++, cnt++) {
 		_engine->_actor->resetActor(a);
@@ -362,7 +362,7 @@ bool Scene::loadSceneLBA1() {
 		}
 	}
 
-	_sceneNumZones = stream.readUint16LE();
+	_sceneNumZones = (int16)stream.readUint16LE();
 	for (int32 i = 0; i < _sceneNumZones; i++) {
 		ZoneStruct *zone = &_sceneZones[i];
 		zone->mins.x = (int16)stream.readUint16LE();
